etl/requirements.txt
pandas==2.2.3
psycopg2-binary==2.9.10
pytest==8.3.3
python-dotenv==1.0.1
requests==2.32.3
sqlalchemy==2.0.36
testing.postgresql==1.3.0